Allow C++ namespacing of tables and end-of-input handling
Previously, it was impossible to have multiple automata implemented
in the same file because all the data tables for an automaton were
defined in the code unit (i.e. file) as `static` linkage constants.  By
using this new option for putting the tables of each parser in a C++
namespace, implementations for multiple automata can live in the same
file.

Because the target language for the generated implementation may not
support a unique value for "end-of-input" to inject as the next
character to process when the end of the input stream is reached --
and therefore, no transition actions are possible -- the `State#at_end`
method is introduced to define actions to be taken depending on the
final state.  It is up to the code generator to decide how end of input
will be notified in order to trigger any specified "at end" actions.
The provided C++ generator creates a `inputEnds()` member function to
be called on the automaton object when the end of input is reached.  If
no `at_end` actions are defined in the automaton, it is optional for
the C++ client code to call `inputEnds()`, but the possibility of
future inclusion of `at_end` actions makes the omission fragile.